moved all feedback prompt fns to agent
