+++ autogpt/app.py
@@ -1,12 +1,11 @@
 """ Command and Control """
 import json
-from typing import Any, Dict, List, NoReturn, Tuple, Union
+from typing import Any, Dict, List, Tuple, Union
 
 from autogpt.agent.agent_manager import AgentManager
 from autogpt.commands.command import CommandRegistry, command
 from autogpt.commands.web_requests import scrape_links, scrape_text
 from autogpt.config import Config
-from autogpt.logs import logger
 from autogpt.memory import get_memory
 from autogpt.processing.text import summarize_text
 from autogpt.prompts.generator import PromptGenerator
