+++ autogpt/prompts/generator.py
@@ -2,9 +2,6 @@
 import json
 from typing import TYPE_CHECKING, Any, Callable, Dict, List, Optional
 
-if TYPE_CHECKING:
-    from autogpt.llm.command_error import CommandError
-
 
 class PromptGenerator:
     """
@@ -110,7 +107,7 @@ class PromptGenerator:
         """
         self.performance_evaluation.append(evaluation)
 
-    def _generate_numbered_list(self, items: List[Any], item_type="list") -> str:
+    def generate_numbered_list(self, items: List[Any], item_type="list") -> str:
         """
         Generate a numbered list from given items based on the item_type.
 
@@ -146,41 +143,13 @@ class PromptGenerator:
         """
         formatted_response_format = json.dumps(self.response_format, indent=4)
         return (
-            f"Constraints:\n{self._generate_numbered_list(self.constraints)}\n\n"
+            f"Constraints:\n{self.generate_numbered_list(self.constraints)}\n\n"
             "Commands:\n"
-            f"{self._generate_numbered_list(self.commands, item_type='command')}\n\n"
-            f"Resources:\n{self._generate_numbered_list(self.resources)}\n\n"
+            f"{self.generate_numbered_list(self.commands, item_type='command')}\n\n"
+            f"Resources:\n{self.generate_numbered_list(self.resources)}\n\n"
             "Performance Evaluation:\n"
-            f"{self._generate_numbered_list(self.performance_evaluation)}\n\n"
+            f"{self.generate_numbered_list(self.performance_evaluation)}\n\n"
             "You should only respond in JSON format as described below \nResponse"
             f" Format: \n{formatted_response_format} \nEnsure the response can be"
             " parsed by Python json.loads"
-        )
-
-    def generate_feedback_prompt_string(self, thoughts: dict[str, str]) -> str:
-        """
-        Generate a prompt string based on the constraints, commands, resources,
-            and performance evaluations.
-
-        Returns:
-            str: The generated prompt string.
-        """
-        thought = thoughts.get("text", "")
-        reasoning = thoughts.get("reasoning", "")
-        plan = thoughts.get("plan", "")
-
-        return (
-            f"Constraints:\n{self._generate_numbered_list(self.constraints)}\n\n"
-            "Commands:\n"
-            f"{self._generate_numbered_list(self.commands, item_type='command')}\n\n"
-            f"Thoughts: {thought}\n"
-            f"Reasoning: {reasoning}\n"
-            f"Plan:\n{plan}\n\n"
-            "You should respond with a concise paragraph that contains any "
-            "improvements to my overall thoughts, reasoning, and plan. Based "
-            "on these improvements, include a list of possible actions "
-            "that will get us closer to our goals, while avoiding previous "
-            "errors and only using the commands provided with valid "
-            "arguments. Order these options from most likely to be effective "
-            "to least."
         )
